MyBot/requirements.txt
streamlit
langchain
langchain-huggingface
langchain-community
langchain-google-genai
langchain-chroma
chromadb
python-dotenv
pandas